POM refactoring
