+++ diplom3/.gitignore
@@ -10,6 +10,7 @@
 # Mobile Tools for Java (J2ME)
 .mtj.tmp/
 /target
+.allure
 *.iml
 # Package Files #
 *.jar
@@ -19,6 +20,7 @@
 *.zip
 *.tar.gz
 *.rar
+build
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
