+++ diplom3/src/test/java/Pages/MainPage.java
@@ -4,36 +4,19 @@ import com.codeborne.selenide.SelenideElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.How;
 
-public class MainPage extends Main {
-    //локатор поля ввода email
-    @FindBy(how = How.NAME, using = "name")
-    private SelenideElement emailField;
-    @FindBy(how = How.XPATH, using = "(//*//input[@name='name'])[2]")
-    private SelenideElement emailRegField;
+import static com.codeborne.selenide.Selenide.page;
+
+public class MainPage {
 
 
-    //локатор поля ввода пароля
-    @FindBy(how = How.NAME, using = "Пароль")
-    private SelenideElement passwordField;
-    //локатор кнопки входа в приложение
-    @FindBy(how = How.XPATH, using = "//div[@id='root']/div/main/div/form/button")
-    private SelenideElement signInButton;
-    //локатор Орормить заказ
-    @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Оформить заказ']/parent::*")
-    private SelenideElement makeOrderButton;
     // Локатор Войти в аккаунт
     @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Войти в аккаунт']/parent::*")
     private SelenideElement signInBody;
-    // Локатор Войти в аккаунт
+    // Локатор Личный кабинет
     @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Личный Кабинет']/parent::*")
     private SelenideElement signInHeader;
-    // Локатор Вход после регистрации
-    @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Вход']/parent::*")
-    private SelenideElement loginInHeader;
 
 
-    @FindBy(how = How.LINK_TEXT, using = "Войти")
-    private SelenideElement signInRegisterPage;
     // Локатор Булки
     @FindBy(how = How.XPATH, using = "(.//*[normalize-space(text()) and normalize-space(.)='Булки'])/parent::*")
     private SelenideElement bunButton;
@@ -54,24 +37,13 @@ public class MainPage extends Main {
 
     @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Профиль']/parent::*")
     private SelenideElement userHeader;
-
-    @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Некорректный пароль']/parent::*")
-    private SelenideElement wrongPasswordError;
-
-    @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Выход']/parent::*")
-    private SelenideElement logoutButton;
-
-    @FindBy(how = How.XPATH, using = "(.//h2[normalize-space(text()) and normalize-space(.)='Вход'])[1]")
-    private SelenideElement loginHeader;
-
     @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Личный Кабинет']/parent::*")
     private SelenideElement dashboardButton;
 
-    @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Конструктор']/parent::*")
-    private SelenideElement constructorButton;
 
-    @FindBy(how = How.XPATH, using = "//*[contains(@class, 'AppHeader_header__logo')]")
-    private SelenideElement headerLogo;
+
+
+
     @FindBy(how = How.XPATH, using = "//*/text()[normalize-space(.)='Соберите бургер']/parent::*")
     private SelenideElement constructYourBurgerHeader;
 
@@ -79,58 +51,32 @@ public class MainPage extends Main {
         return constructYourBurgerHeader.getText();
     }
 
-    public void clickHeaderLogo() {
-        headerLogo.click();
-    }
+
 
     //метод клика по ссылке авторизации
-    public void clickSignInBody() {
+    public LoginPage clickSignInBody() {
         signInBody.click();
+        return page(LoginPage.class);
     }
 
-    //метод клика по ссылке авторизации
-    public void clickSignInHeader() {
+    //метод клика по кнопке Личный кабинет
+    public DashboardPage clickSignInHeader() {
         signInHeader.click();
+        return page(DashboardPage.class);
     }
 
-    //метод клика по кнопке Конструктор
-    public void clickConstructorButton() {
-        constructorButton.click();
+    public LoginPage clickSignInHeaderLogin() {
+        signInHeader.click();
+        return page(LoginPage.class);
     }
 
-    //метод клика по ссылке авторизации
-    public void clickSignInRegPage() {
-        signInRegisterPage.click();
-    }
 
-    //метод заполнения поля ввода email
-    public void setEmail(String username) {
-        emailField.setValue(username);
-    }
 
-    public void setRegEmail(String username) {
-        emailRegField.setValue(username);
-    }
 
-    public void setName(String name) {
-        emailField.setValue(name);
+    public DashboardPage dashboardButtonClick() {
+        dashboardButton.click();
+        return page(DashboardPage.class);
     }
-
-    //метод заполнения поля ввода пароля
-    public void setPassword(String password) {
-        passwordField.setValue(password);
-    }
-
-    //метод клика по кнопке авторизации
-    public void clickSignInButton() {
-        signInButton.click();
-    }
-
-    //метод получить текст
-    public String getMakeOrderText() {
-        return makeOrderButton.getText();
-    }
-
 
     public void clickBunButton() {
         bunButton.click();
@@ -146,16 +92,6 @@ public class MainPage extends Main {
         fillersButton.click();
     }
 
-    public void scrollToElement() {
-        fillersHeader.scrollTo();
-    }
-
-    //метод авторизации в приложении: объединяет ввод email, пароля и клик по кнопке
-    public void login(String username, String password) {
-        setEmail(username);
-        setPassword(password);
-        clickSignInButton();
-    }
 
     public String getBunHeader() {
         return bunHeader.getText();
@@ -169,41 +105,5 @@ public class MainPage extends Main {
         return sauceHeader.getText();
     }
 
-    public String getUser() {
-        return userHeader.getText();
-    }
 
-    public String getWrongPasswordTest() {
-        return wrongPasswordError.getText();
-    }
-
-    public String getLoginInHeader() {
-        return loginInHeader.getText();
-    }
-
-
-    public void register(String name, String email, String password) {
-        setName(name);
-        setRegEmail(email);
-        setPassword(password);
-        clickSignInButton();
-    }
-
-
-    public void dashboardButtonClick() {
-        dashboardButton.click();
-    }
-
-    public void logOutButtonClick() {
-        logoutButton.click();
-    }
-
-    public void logOut() {
-        dashboardButtonClick();
-        logOutButtonClick();
-    }
-
-    public String getLoginHeader() {
-        return loginHeader.getText();
-    }
 }
